oh wait
this little duck didnt catch up fast enough
the train left without it
it had to catch the next one
it later found the rest of its family, which had forgotten it and moved
on with their lives, not even recognizing it.
falling deep into depression, this little duck decided to work on a
spaceship to put lazers into space and get revenge on it's unloving
family.
"They'll see... I'll show them..."